Move ptr_escape tests to a sibling file
The inline test module was more than half of ptr_escape.rs. Also point
the PtrEscapeSummary doc at the production call-site derivation path
(escape_scan) instead of a test-only helper.
